Added general button class
Created the button class set up.

We said we wanted to have two classes and I said I would take care of the ImageButton class as well but I feel I wanted to nail the button class first.

Missing from the task list here is the set up for a redraw method.  I figured I could go over that more and just finish it up pretty quick before moving on to the other class which will also be housed in this file, most likely through inheritance.

Lastly, for this class we set some attributes for the object those being the self.image which can be resized through the scale parameter.  Self.rect is created by using the pygame get_rect() method and returns the rectangle we'll use to create a rectangle object out of the image.  The rect.topleft attribute is set through the x and y coordinates.

The draw method simply draws the image onto the screen taking the image and x and y coordinates as arguments